test(installer): promote the stub-server test vehicle to a shared module (#4246)
The only real test vehicle in this crate — a raw `tokio::net::TcpListener`
answering fixed HTTP responses plus a `TRUSTY_DATA_DIR_OVERRIDE`-stubbed data
dir with a real `http_addr` planted in it — lived inline in
`ensure::project_setup`'s test module. #4246 needs that same vehicle from
`probe_http` and `verify_tail`, and `trusty-installer` has an empty
`[dev-dependencies]` (no wiremock, and adding one is out of scope).

Promote it to `commands::test_support`, move the process-global `ENV_TEST_LOCK`
there (re-exported from `ensure` so the three existing importers are untouched),
and fold the repeated set_var/remove_var/remove_dir_all teardown into
`stub_empty_data_dir` / `clear_data_dir_override` so the `unsafe` justification
is argued once rather than ten times.

No production code changes. 483 tests still pass.
